Changed some of the data sent from mcs_send_connection_initial - the number of channels being the most important.
Don't assume the length of the data coming back from the server is
of the same length that the data we sent was.

Record the channel id of incoming MCS packets.

Send username to iso_connect, in order for it to be able to send the
mstshash.

Open the clipboard channel if we are speaking RDP5.


git-svn-id: svn://svn.code.sf.net/p/rdesktop/code/trunk/rdesktop@347 6ed3e302-cbb0-46be-b2a8-fc98bdbf2350