Send different logon packets based on if we try to support RDP5 or not. The order of the data is slightly different in RDP5.
Moved pointer PDU handling to their own functions, since RDP5 no longer
encapsulates ColorPointer and CachedPointer in a PDU with a type field.

Made debug output when doing Bitmap updates clearer. Also added debug
printout when bitmap compression failed.

Added code that checks for a new way of sending the size of the compressed
bitmaps. Code imported from Matt's RDP5 adventures :).

Debug output when doing Palette updates in order to make it easier to follow
packets.

Fetch the number of orders here instead of in process_orders and send it
along to that function as an argument.

Send the username to sec_connect.


git-svn-id: svn://svn.code.sf.net/p/rdesktop/code/trunk/rdesktop@350 6ed3e302-cbb0-46be-b2a8-fc98bdbf2350