Remove process of packets in rdp_connect.
The process of packets in rdp_connect assumes that the
first packets are demand active to set g_rdp_shareid which
is not true. In case when CredSSP+Kerberos is in use an
enhanced redirection PDU packet is sent before the demande
active PDU.

This fixes the problem with rdesktop silently exists in
the case above.



git-svn-id: svn://svn.code.sf.net/p/rdesktop/code/rdesktop/trunk@1794 6ed3e302-cbb0-46be-b2a8-fc98bdbf2350